+++ HISTORY.md
@@ -1,3 +1,10 @@
+# Release 0.5.1
+
+This update adds new variational inference algorithms in light of the flexibility added in the v0.5 update.
+Specifically, the following measure-space optimization algorithms have been added:
+
+  - `KLMinWassFwdBwd`
+
 # Release 0.5
 
 ## Default Configuration Changes